move main camera program to main.cpp
